refactor: vastly improve the DX of src/ui
