+++ src/ui/components/mod.rs
@@ -1,0 +1,24 @@
+pub mod action_bar;
+pub mod command_bar;
+pub mod elements;
+pub mod footer;
+pub mod header;
+pub mod menu;
+pub mod modal;
+pub mod mode_bar;
+pub mod results;
+pub mod size_warning;
+pub mod typing_area;
+
+// re-exports
+pub use action_bar::ActionBarComponent;
+pub use command_bar::CommandBarComponent;
+pub use elements::TermiElement;
+pub use footer::FooterComponent;
+pub use header::HeaderComponent;
+pub use menu::MenuComponent;
+pub use modal::ModalComponent;
+pub use mode_bar::ModeBarComponent;
+pub use results::ResultsComponent;
+pub use size_warning::SizeWarningComponent;
+pub use typing_area::TypingAreaComponent;
