--- src/ui/ui.rs
@@ -1,6 +1,0 @@
-pub use self::render::draw_ui;
-
-pub mod elements;
-pub mod layout;
-pub mod render;
-pub mod utils;
